Style hero image to be samller on smaller devices
Afer presenting the website to a user on a mobile device it appeared
different then how it does in Google chromes dev tools. Google chromes
dev tool doesn't display the browser at the top so the homepage header
is visible. On a real mobile however when the user lands on the page the
browser pushes the hero image down causing the homepage header not to be
visiable meaning the user doesn't get a glims of what else is on the
page and not encouraged to scroll down.