Style element to adapt to language changes
Elements and their childs styled in this commit are as follows:

body - to stop the google translate banner being displayed.
site-header - so the navbar looks like part of the header when its
pushed down. And to stop the last language being a different colour.
due to it being affected by the same selector used to style the
book-now list item in the mobile menu.
hero - to keep the hero text inside the hero image.
button and cta - to keep the button text inside the button.
product-card - to keep its content inside the card.
reviews - to stop the carousel indicators label wrapping to a new line.