book-now: Add content to page and style it
Provides a form for the user to book a lesson and for the client to get
the users information to process the booking.

Style form to be different from other form on the website so it stands
out. The more professional the form looks the more trustworthy the user
will find the website. Style the booking form to look professional so
the user feels comfortable providing their details.

Form set up to post data to https://formdump.codeinstitute.net/ for
testing purposes.