Add booking payment breadcrumb navbar
Informs the user which stage they are on in the booking process.

Currently there is no legitimate way to reach the booking confirmation
or payment page other than by typing in the url.
Due to no backend set up to process the checkout.
An unsupported method ('post') error needs fixing to submit the booking
from to progress to the payment page.
On the payment page the breadcrumb__item for the booking confirmation
page is linked so the the booking confirmation page can be view.